+++ ChangeLog.md
@@ -1,3 +1,7 @@
 # Changelog for easy-imports
+0.1.0.1
+- ~~update cabal file from `easy-imports` automatically~~ (added a flag, because of the speed using stack)
+- updateCabal (updates cabal file from stack file automatically)
 
 ## Unreleased changes
+